+++ packages/commerce-sdk-react/src/auth/index.ts
@@ -38,7 +38,8 @@ interface AuthConfig extends ApiClientConfigParams {
     clientSecret?: string
     silenceWarnings?: boolean
     logger: Logger
-    cookieTtl?: number
+    expirationTimeGuestToken?: number
+    expirationTimeRegisteredToken?: number
 }
 
 interface JWTHeaders {
@@ -179,7 +180,8 @@ class Auth {
     private clientSecret: string
     private silenceWarnings: boolean
     private logger: Logger
-    private cookieTtl: number | undefined
+    private expirationTimeGuestToken: number | undefined
+    private expirationTimeRegisteredToken: number | undefined
 
     constructor(config: AuthConfig) {
         // Special endpoint for injecting SLAS private client secret.
@@ -231,7 +233,8 @@ class Auth {
 
         // TODO: add check where if user increases cookie TTL past acceptable value
         // we log a warning and set it to default
-        this.cookieTtl = config.cookieTtl
+        this.expirationTimeGuestToken = config.expirationTimeGuestToken
+        this.expirationTimeRegisteredToken = config.expirationTimeRegisteredToken
 
         /*
          * There are 2 ways to enable SLAS private client mode.
@@ -383,10 +386,13 @@ class Auth {
         this.set('customer_type', isGuest ? 'guest' : 'registered')
 
         const refreshTokenKey = isGuest ? 'refresh_token_guest' : 'refresh_token_registered'
+        const expirationTime = isGuest
+            ? this.expirationTimeGuestToken
+            : this.expirationTimeRegisteredToken
 
         this.set(refreshTokenKey, res.refresh_token, {
             // TODO: This is where we set this expiration
-            expires: this.cookieTtl ?? res.refresh_token_expires_in
+            expires: expirationTime ?? res.refresh_token_expires_in
         })
     }