+++ packages/commerce-sdk-react/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v3.1.0-dev (Aug 08, 2024)
+- Allow customization of expiration time for guest/registered user refresh tokens [#1977](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1977)
 
 ## v3.0.0 (Aug 07, 2024)
 - Add `meta.displayName` to queries. It can be used to identify queries in performance metrics or logs. [#1895](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1895)